# openapi-mvp-v2.yaml
openapi: 3.0.3
info:
  title: Nomad API (MVP v2)
  version: 2.0.0
  description: |
    Expanded MVP API contract for Nomad with full hook coverage:
    - Auth + Onboarding restored
    - CRUD-expanded resources for courses, communities, posts, and messaging
servers:
  - url: http://localhost:3000/api
    description: Local dev
  - url: https://api.nomad.example.com
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: List courses
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: department
          schema: { type: string }
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [Intro, Intermediate, Advanced, Graduate]
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCourses" }
    post:
      tags: [Courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourseRequest" }
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
  /courses/{courseId}:
    get:
      tags: [Courses]
      summary: Get course details
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
    patch:
      tags: [Courses]
      summary: Update course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCourseRequest" }
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
    delete:
      tags: [Courses]
      summary: Delete course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Course deleted
  /courses/{courseId}/students:
    get:
      tags: [Courses]
      summary: Get students in course
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - in: query
          name: connected
          schema: { type: boolean }
      responses:
        "200":
          description: Classmates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedUsers" }
  /courses/{courseId}/connections:
    get:
      tags: [Courses]
      summary: Get known vs suggested classmates
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Connection buckets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseConnections" }
  /planner/courses:
    get:
      tags: [Planner]
      summary: Get planner state
      responses:
        "200":
          description: Planner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlannerState" }
    put:
      tags: [Planner]
      summary: Replace planner course IDs
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePlannerRequest" }
      responses:
        "200":
          description: Updated planner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlannerState" }
  /communities:
    get:
      tags: [Communities]
      summary: List communities
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: joined
          schema: { type: boolean }
      responses:
        "200":
          description: Communities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCommunities" }
    post:
      tags: [Communities]
      summary: Create community
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommunityRequest" }
      responses:
        "201":
          description: Created community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
  /communities/{communityId}:
    get:
      tags: [Communities]
      summary: Get community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "200":
          description: Community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
    patch:
      tags: [Communities]
      summary: Update community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCommunityRequest" }
      responses:
        "200":
          description: Updated community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
    delete:
      tags: [Communities]
      summary: Delete community
      parameters:
        - $ref: "#/components/parameters/CommunityId"
      responses:
        "204":
          description: Community deleted
...
  /posts:
    get:
      tags: [Feed]
      summary: List posts
      parameters:
        - in: query
          name: tab
          schema:
            type: string
            enum: [all, following, top]
        - in: query
          name: communityId
          schema: { type: string }
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPosts" }
    post:
      tags: [Feed]
      summary: Create post
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
  /posts/{postId}:
    patch:
      tags: [Feed]
      summary: Update post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePostRequest" }
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
    delete:
      tags: [Feed]
      summary: Delete post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Post deleted
  /posts/{postId}/reactions:
    post:
      tags: [Feed]
      summary: React to post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostReactionRequest" }
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Post" }
  /posts/{postId}/replies:
    post:
      tags: [Feed]
      summary: Create reply
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReplyRequest" }
      responses:
        "201":
          description: Reply created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reply" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CourseId:
      in: path
      name: courseId
      required: true
      schema: { type: string }
    CommunityId:
      in: path
      name: communityId
      required: true
      schema: { type: string }
    PostId:
      in: path
      name: postId
      required: true
      schema: { type: string }
...
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatar: { type: string }
        university: { type: string }
        major: { type: string }
        minor: { type: string, nullable: true }
        year: { type: string }
        bio: { type: string }
        headline: { type: string, nullable: true }
        interests:
          type: array
          items: { type: string }
        courses:
          type: array
          items: { type: string }
        communities:
          type: array
          items: { type: string }
        isConnected: { type: boolean }
        isOnline: { type: boolean }
...
    Course:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        title: { type: string }
        credits: { type: integer }
        description: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        nextCourses:
          type: array
          items: { type: string }
        department: { type: string }
        difficulty:
          type: string
          enum: [Intro, Intermediate, Advanced, Graduate]
        tags:
          type: array
          items: { type: string }
        students:
          type: array
          items: { type: string }
        rating: { type: number, format: float }
    CreateCourseRequest:
      type: object
      properties:
        code: { type: string }
        title: { type: string }
        credits: { type: integer }
        description: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        nextCourses:
          type: array
          items: { type: string }
        department: { type: string }
        difficulty:
          type: string
          enum: [Intro, Intermediate, Advanced, Graduate]
        tags:
          type: array
          items: { type: string }
    UpdateCourseRequest:
      allOf:
        - $ref: "#/components/schemas/CreateCourseRequest"
    PlannerState:
      type: object
      properties:
        courseIds:
          type: array
          items: { type: string }
    UpdatePlannerRequest:
      type: object
      required: [courseIds]
      properties:
        courseIds:
          type: array
          items: { type: string }
    Community:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string, nullable: true }
        color: { type: string }
        members: { type: integer }
        posts: { type: integer }
        tags:
          type: array
          items: { type: string }
        isJoined: { type: boolean }
        university: { type: string, nullable: true }
    CreateCommunityRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string }
        color: { type: string }
        tags:
          type: array
          items: { type: string }
        university: { type: string }
    UpdateCommunityRequest:
      allOf:
        - $ref: "#/components/schemas/CreateCommunityRequest"
...
    Post:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        communityId: { type: string }
        title: { type: string }
        content: { type: string }
        timestamp: { type: string, format: date-time }
        likes: { type: integer }
        isLiked: { type: boolean }
        tags:
          type: array
          items: { type: string }
        replies:
          type: array
          items: { $ref: "#/components/schemas/Reply" }
    Reply:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        content: { type: string }
        timestamp: { type: string, format: date-time }
        likes: { type: integer }
        isLiked: { type: boolean }
    CreatePostRequest:
      type: object
      required: [communityId, title, content]
      properties:
        communityId: { type: string }
        title: { type: string }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
    UpdatePostRequest:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
    PostReactionRequest:
      type: object
      required: [reaction]
      properties:
        reaction:
          type: string
          enum: [like, celebrate, insightful, curious, support]
    CreateReplyRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
...
    CourseConnections:
      type: object
      properties:
        known:
          type: array
          items: { $ref: "#/components/schemas/User" }
        suggested:
          type: array
          items: { $ref: "#/components/schemas/User" }
...
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        total: { type: integer }
    PaginatedCourses:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Course" }
        total: { type: integer }
    PaginatedCommunities:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Community" }
        total: { type: integer }
    PaginatedPosts:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Post" }
        total: { type: integer }
...